Allow build of nullaway with JDK21
PreservedAnnotationTreeVisitor.JDK17AndEarlierTypeMetadataBuilder
had to be built with JDK 17, or an earlier version. It was the
case that building with JDK 21 would cause build errors due to
changes in com.sun.tools.javac.code.TypeMetadata.Annotations and
com.sun.tools.javac.code.Type.cloneWithMetadata().

This change replicates the usage of MethodHandles used in
JDK21AndEarlierTypeMetadataBuilder to avoid compile time
references to the changed methods.

Bug: 313924276
Test: TH
Change-Id: I5bd38c66ebc058fc1fa206752a18de5412cf80fe